FIX: Minor fixes.

+++ lib/common/rootfs_script.sh
@@ -35,7 +35,7 @@ function locale_setup_voidlinux() {
 
 function grub_config() {
   if [[ $bootloader_type = uefi ]]; then
-    grub-install --target=i386-efi --efi-directory=/boot --removable $grub_param
+    [[ -f /usr/lib/grub/i386-efi/modinfo.sh ]] && grub-install --target=i386-efi --efi-directory=/boot --removable $grub_param
     grub-install --target=x86_64-efi --efi-directory=/boot --removable $grub_param
   else
     grub-install --target=i386-pc --force $grub_param $bootloader_bios_place
